@@ -5,7 +5,6 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"path/filepath"
@@ -44,6 +43,7 @@ type PublishedRepo struct {
 	ButAutomaticUpgrades string
 	Label                string
 	Suite                string
+	Codename             string
 	// Architectures is a list of all architectures published
 	Architectures []string
 	// SourceKind is "local"/"repo"
@@ -61,6 +61,9 @@ type PublishedRepo struct {
 
 	// Skip contents generation
 	SkipContents bool
+
+	// Skip bz2 compression for index files
+	SkipBz2 bool
 
 	// True if repo is being re-published
 	rePublishing bool
@@ -309,6 +312,7 @@ func (p *PublishedRepo) MarshalJSON() ([]byte, error) {
 		"Label":                p.Label,
 		"Origin":               p.Origin,
 		"Suite":                p.Suite,
+		"Codename":             p.Codename,
 		"NotAutomatic":         p.NotAutomatic,
 		"ButAutomaticUpgrades": p.ButAutomaticUpgrades,
 		"Prefix":               p.Prefix,
@@ -361,6 +365,10 @@ func (p *PublishedRepo) String() string {
 
 	if p.Suite != "" {
 		extras = append(extras, fmt.Sprintf("suite: %s", p.Suite))
+	}
+
+	if p.Codename != "" {
+		extras = append(extras, fmt.Sprintf("codename: %s", p.Codename))
 	}
 
 	extra = strings.Join(extras, ", ")
@@ -491,7 +499,7 @@ func (p *PublishedRepo) GetLabel() string {
 	return p.Label
 }
 
-// GetName returns the unique name of the repo
+// GetPath returns the unique name of the repo
 func (p *PublishedRepo) GetPath() string {
 	prefix := p.StoragePrefix()
 
@@ -508,6 +516,14 @@ func (p *PublishedRepo) GetSuite() string {
 		return p.Distribution
 	}
 	return p.Suite
+}
+
+// GetCodename returns default or manual Codename:
+func (p *PublishedRepo) GetCodename() string {
+	if p.Codename == "" {
+		return p.Distribution
+	}
+	return p.Codename
 }
 
 // Publish publishes snapshot (repository) contents, links package files, generates Packages & Release files, signs them
@@ -579,13 +595,13 @@ func (p *PublishedRepo) Publish(packagePool aptly.PackagePool, publishedStorageP
 	}
 
 	var tempDir string
-	tempDir, err = ioutil.TempDir(os.TempDir(), "aptly")
+	tempDir, err = os.MkdirTemp(os.TempDir(), "aptly")
 	if err != nil {
 		return err
 	}
 	defer os.RemoveAll(tempDir)
 
-	indexes := newIndexFiles(publishedStorage, basePath, tempDir, suffix, p.AcquireByHash)
+	indexes := newIndexFiles(publishedStorage, basePath, tempDir, suffix, p.AcquireByHash, p.SkipBz2)
 
 	legacyContentIndexes := map[string]*ContentsIndex{}
 	var count int64
@@ -732,6 +748,7 @@ func (p *PublishedRepo) Publish(packagePool aptly.PackagePool, publishedStorageP
 				release["Origin"] = p.GetOrigin()
 				release["Label"] = p.GetLabel()
 				release["Suite"] = p.GetSuite()
+				release["Codename"] = p.GetCodename()
 				if p.AcquireByHash {
 					release["Acquire-By-Hash"] = "yes"
 				}
@@ -790,7 +807,7 @@ func (p *PublishedRepo) Publish(packagePool aptly.PackagePool, publishedStorageP
 	}
 	release["Label"] = p.GetLabel()
 	release["Suite"] = p.GetSuite()
-	release["Codename"] = p.Distribution
+	release["Codename"] = p.GetCodename()
 	release["Date"] = time.Now().UTC().Format("Mon, 2 Jan 2006 15:04:05 MST")
 	release["Architectures"] = strings.Join(utils.StrSlicesSubstract(p.Architectures, []string{ArchitectureSource}), " ")
 	if p.AcquireByHash {
