@@ -3,7 +3,6 @@ package http
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
@@ -26,7 +25,7 @@ type GrabDownloaderSuiteBase struct {
 }
 
 func (s *GrabDownloaderSuiteBase) SetUpTest(c *C) {
-	s.tempfile, _ = ioutil.TempFile(os.TempDir(), "aptly-test")
+	s.tempfile, _ = os.CreateTemp(os.TempDir(), "aptly-test")
 	s.l, _ = net.ListenTCP("tcp4", &net.TCPAddr{IP: net.IPv4(127, 0, 0, 1)})
 	s.url = fmt.Sprintf("http://localhost:%d", s.l.Addr().(*net.TCPAddr).Port)
 
@@ -42,7 +41,7 @@ func (s *GrabDownloaderSuiteBase) SetUpTest(c *C) {
 		close(s.ch)
 	}()
 
-	s.progress = console.NewProgress()
+	s.progress = console.NewProgress(false)
 	s.progress.Start()
 
 	s.d = NewGrabDownloader(0, 1, s.progress)
