@@ -71,7 +71,7 @@ func apiPublishList(c *gin.Context) {
 	})
 
 	if err != nil {
-		c.AbortWithError(500, err)
+		AbortWithJSONError(c, 500, err)
 		return
 	}
 
@@ -96,6 +96,7 @@ func apiPublishRepoOrSnapshot(c *gin.Context) {
 		ButAutomaticUpgrades string
 		ForceOverwrite       bool
 		SkipContents         *bool
+		SkipBz2              *bool
 		Architectures        []string
 		Signing              SigningOptions
 		AcquireByHash        *bool
@@ -107,12 +108,12 @@ func apiPublishRepoOrSnapshot(c *gin.Context) {
 
 	signer, err := getSigner(&b.Signing)
 	if err != nil {
-		c.AbortWithError(500, fmt.Errorf("unable to initialize GPG signer: %s", err))
+		AbortWithJSONError(c, 500, fmt.Errorf("unable to initialize GPG signer: %s", err))
 		return
 	}
 
 	if len(b.Sources) == 0 {
-		c.AbortWithError(400, fmt.Errorf("unable to publish: soures are empty"))
+		AbortWithJSONError(c, 400, fmt.Errorf("unable to publish: soures are empty"))
 		return
 	}
 
@@ -133,14 +134,14 @@ func apiPublishRepoOrSnapshot(c *gin.Context) {
 
 			snapshot, err = snapshotCollection.ByName(source.Name)
 			if err != nil {
-				c.AbortWithError(404, fmt.Errorf("unable to publish: %s", err))
+				AbortWithJSONError(c, 404, fmt.Errorf("unable to publish: %s", err))
 				return
 			}
 
 			resources = append(resources, string(snapshot.ResourceKey()))
 			err = snapshotCollection.LoadComplete(snapshot)
 			if err != nil {
-				c.AbortWithError(500, fmt.Errorf("unable to publish: %s", err))
+				AbortWithJSONError(c, 500, fmt.Errorf("unable to publish: %s", err))
 				return
 			}
 
@@ -157,26 +158,26 @@ func apiPublishRepoOrSnapshot(c *gin.Context) {
 
 			localRepo, err = localCollection.ByName(source.Name)
 			if err != nil {
-				c.AbortWithError(404, fmt.Errorf("unable to publish: %s", err))
+				AbortWithJSONError(c, 404, fmt.Errorf("unable to publish: %s", err))
 				return
 			}
 
 			resources = append(resources, string(localRepo.Key()))
 			err = localCollection.LoadComplete(localRepo)
 			if err != nil {
-				c.AbortWithError(500, fmt.Errorf("unable to publish: %s", err))
+				AbortWithJSONError(c, 500, fmt.Errorf("unable to publish: %s", err))
 			}
 
 			sources = append(sources, localRepo)
 		}
 	} else {
-		c.AbortWithError(400, fmt.Errorf("unknown SourceKind"))
+		AbortWithJSONError(c, 400, fmt.Errorf("unknown SourceKind"))
 		return
 	}
 
 	published, err := deb.NewPublishedRepo(storage, prefix, b.Distribution, b.Architectures, components, sources, collectionFactory)
 	if err != nil {
-		c.AbortWithError(500, fmt.Errorf("unable to publish: %s", err))
+		AbortWithJSONError(c, 500, fmt.Errorf("unable to publish: %s", err))
 		return
 	}
 
@@ -207,6 +208,11 @@ func apiPublishRepoOrSnapshot(c *gin.Context) {
 		published.SkipContents = context.Config().SkipContentsPublishing
 		if b.SkipContents != nil {
 			published.SkipContents = *b.SkipContents
+		}
+
+		published.SkipBz2 = context.Config().SkipBz2Publishing
+		if b.SkipBz2 != nil {
+			published.SkipBz2 = *b.SkipBz2
 		}
 
 		if b.AcquireByHash != nil {
@@ -243,6 +249,7 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 		ForceOverwrite bool
 		Signing        SigningOptions
 		SkipContents   *bool
+		SkipBz2        *bool
 		SkipCleanup    *bool
 		Snapshots      []struct {
 			Component string `binding:"required"`
@@ -257,7 +264,7 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 
 	signer, err := getSigner(&b.Signing)
 	if err != nil {
-		c.AbortWithError(500, fmt.Errorf("unable to initialize GPG signer: %s", err))
+		AbortWithJSONError(c, 500, fmt.Errorf("unable to initialize GPG signer: %s", err))
 		return
 	}
 
@@ -266,12 +273,12 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 
 	published, err := collection.ByStoragePrefixDistribution(storage, prefix, distribution)
 	if err != nil {
-		c.AbortWithError(404, fmt.Errorf("unable to update: %s", err))
+		AbortWithJSONError(c, 404, fmt.Errorf("unable to update: %s", err))
 		return
 	}
 	err = collection.LoadComplete(published, collectionFactory)
 	if err != nil {
-		c.AbortWithError(500, fmt.Errorf("unable to update: %s", err))
+		AbortWithJSONError(c, 500, fmt.Errorf("unable to update: %s", err))
 		return
 	}
 
@@ -281,7 +288,7 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 
 	if published.SourceKind == deb.SourceLocalRepo {
 		if len(b.Snapshots) > 0 {
-			c.AbortWithError(400, fmt.Errorf("snapshots shouldn't be given when updating local repo"))
+			AbortWithJSONError(c, 400, fmt.Errorf("snapshots shouldn't be given when updating local repo"))
 			return
 		}
 		updatedComponents = published.Components()
@@ -292,20 +299,20 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 		publishedComponents := published.Components()
 		for _, snapshotInfo := range b.Snapshots {
 			if !utils.StrSliceHasItem(publishedComponents, snapshotInfo.Component) {
-				c.AbortWithError(404, fmt.Errorf("component %s is not in published repository", snapshotInfo.Component))
+				AbortWithJSONError(c, 404, fmt.Errorf("component %s is not in published repository", snapshotInfo.Component))
 				return
 			}
 
 			snapshotCollection := collectionFactory.SnapshotCollection()
 			snapshot, err2 := snapshotCollection.ByName(snapshotInfo.Name)
 			if err2 != nil {
-				c.AbortWithError(404, err2)
+				AbortWithJSONError(c, 404, err2)
 				return
 			}
 
 			err2 = snapshotCollection.LoadComplete(snapshot)
 			if err2 != nil {
-				c.AbortWithError(500, err2)
+				AbortWithJSONError(c, 500, err2)
 				return
 			}
 
@@ -314,12 +321,16 @@ func apiPublishUpdateSwitch(c *gin.Context) {
 			updatedSnapshots = append(updatedSnapshots, snapshot.Name)
 		}
 	} else {
-		c.AbortWithError(500, fmt.Errorf("unknown published repository type"))
+		AbortWithJSONError(c, 500, fmt.Errorf("unknown published repository type"))
 		return
 	}
 
 	if b.SkipContents != nil {
 		published.SkipContents = *b.SkipContents
+	}
+
+	if b.SkipBz2 != nil {
+		published.SkipBz2 = *b.SkipBz2
 	}
 
 	if b.AcquireByHash != nil {
@@ -365,7 +376,7 @@ func apiPublishDrop(c *gin.Context) {
 
 	published, err := collection.ByStoragePrefixDistribution(storage, prefix, distribution)
 	if err != nil {
-		c.AbortWithError(http.StatusInternalServerError, fmt.Errorf("unable to drop: %s", err))
+		AbortWithJSONError(c, http.StatusNotFound, fmt.Errorf("unable to drop: %s", err))
 		return
 	}
 
