@@ -166,7 +166,7 @@ func isMultilineField(field string, isRelease bool) bool {
 
 // Write single field from Stanza to writer.
 //
-//nolint: interfacer
+// nolint: interfacer
 func writeField(w *bufio.Writer, field, value string, isRelease bool) (err error) {
 	if !isMultilineField(field, isRelease) {
 		_, err = w.WriteString(field + ": " + value + "\n")
