@@ -47,6 +47,9 @@ func (s *ConfigSuite) TestSaveConfig(c *C) {
 	s.config.AzurePublishRoots = map[string]AzurePublishRoot{"test": {
 		Container: "repo"}}
 
+	s.config.LogLevel = "info"
+	s.config.LogFormat = "json"
+
 	err := SaveConfig(configname, &s.config)
 	c.Assert(err, IsNil)
 
@@ -79,6 +82,7 @@ func (s *ConfigSuite) TestSaveConfig(c *C) {
 		"  \"ppaDistributorID\": \"\",\n"+
 		"  \"ppaCodename\": \"\",\n"+
 		"  \"skipContentsPublishing\": false,\n"+
+		"  \"skipBz2Publishing\": false,\n"+
 		"  \"FileSystemPublishEndpoints\": {\n"+
 		"    \"test\": {\n"+
 		"      \"rootDir\": \"/opt/aptly-publish\",\n"+
@@ -101,6 +105,7 @@ func (s *ConfigSuite) TestSaveConfig(c *C) {
 		"      \"plusWorkaround\": false,\n"+
 		"      \"disableMultiDel\": false,\n"+
 		"      \"forceSigV2\": false,\n"+
+		"      \"forceVirtualHostedStyle\": false,\n"+
 		"      \"debug\": false\n"+
 		"    }\n"+
 		"  },\n"+
@@ -129,7 +134,9 @@ func (s *ConfigSuite) TestSaveConfig(c *C) {
 		"    }\n"+
 		"  },\n"+
 		"  \"AsyncAPI\": false,\n"+
-		"  \"enableMetricsEndpoint\": false\n"+
+		"  \"enableMetricsEndpoint\": false,\n"+
+		"  \"logLevel\": \"info\",\n"+
+		"  \"logFormat\": \"json\"\n"+
 		"}")
 }
 
